+++ src/mml/core/data_loading/file_manager.py
@@ -14,7 +14,6 @@ from typing import Dict, List, Optional, Union
 import ijson
 import orjson
 import torch
-from lightning.pytorch.callbacks import ModelCheckpoint
 from omegaconf import Container, DictConfig
 
 from mml.core.data_loading.task_description import (
@@ -40,7 +39,7 @@ REUSABLE_NUMBER_TAG = "#"
 # default strategies to determine file paths for certain computed artefacts
 # key : (type, path, enable_numbering, reusable)
 DEFAULT_ASSIGNMENTS = {
-    "parameters": (ModelCheckpoint, Path("PROJ_PATH") / "PARAMETERS" / "TASK_NAME" / "model.pth", True, True),
+    "parameters": (torch.nn.Module, Path("PROJ_PATH") / "PARAMETERS" / "TASK_NAME" / "model.mml", True, True),
     "img_examples": (None, Path("PROJ_PATH") / "IMG_EXAMPLES" / "TASK_NAME" / "examples.png", True, False),
     "blueprint": (Container, Path("PROJ_PATH") / "BLUEPRINTS" / "TASK_NAME" / "blueprint.yaml", True, True),
     "pipeline": (Container, Path("PROJ_PATH") / "PIPELINES" / "TASK_NAME" / "pipeline.yaml", True, True),
@@ -53,19 +52,27 @@ DEFAULT_ASSIGNMENTS = {
 }
 
 
-# configuration or reusing certain artefacts,
-# during runtime this is an OmegaConf DictConfig and may contain additional entries
 @dataclasses.dataclass
 class ReuseConfig:
+    """
+    Configuration for reusing certain artefacts from other projects or previous runs.
+    During runtime this is an OmegaConf DictConfig and may contain additional entries!
+    See :meth:`_find_reusables` on how this is used.
+    """
+
     blueprint: Optional[Union[str, List[str]]] = None
     models: Optional[Union[str, List[str]]] = None
     parameters: Optional[Union[str, List[str]]] = None
 
 
-# configuration for removing certain artefacts,
-# during runtime this is an OmegaConf DictConfig and may contain additional entries
 @dataclasses.dataclass
 class RemoveConfig:
+    """
+    Configuration for removing certain artefacts from at the end of the run.
+    During runtime this is an OmegaConf DictConfig and may contain additional entries!
+    See :meth:`remove_intermediates` on how this is used.
+    """
+
     img_examples: bool = False
     blueprint: bool = False
     parameters: bool = False
@@ -221,6 +228,7 @@ class MMLFileManager(Singleton):
                 f"Dataset {dset_name} already present at {candidate}! Please choose a different name."
             )
         candidate.mkdir(parents=False, exist_ok=False)
+        logger.debug(f"Created RAW folder for dataset {dset_name}.")
         return candidate
 
     def get_all_dset_names(self) -> Dict[str, Dict[str, Path]]: